LayoutTests/platform/chromium/fast/forms/suggestion-picker/datetime-suggestion-picker-min-max-attribute.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>
</head>
<body style="background-color: #bbbbbb;">
<p id="description"></p>
<div id="console"></div>
<input type=datetime id=datetime value="2012-11-20T12:00Z" min="2012-11-20T06:00Z" max="2012-11-21T12:00Z" list=suggestions>
<datalist id=suggestions>
    <option>2012-11-19T18:00Z</option>
    <option>2012-11-20T00:00Z</option>
    <option>2012-11-20T06:00Z</option>
    <option>2012-11-20T12:00Z</option>
    <option>2012-11-20T18:00Z</option>
    <option>2012-11-21T00:00Z</option>
    <option>2012-11-21T06:00Z</option>
    <option>2012-11-21T12:00Z</option>
    <option>2012-11-21T18:00Z</option>
    <option>2012-11-22T00:00Z</option>
</datalist>
<script>
description('Tests that min/max attributes filter out suggestions as expected.');

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

window.onload = function() {
    openPicker(document.getElementById('datetime'), test);
};

function test() {
    popupWindow.focus();

    shouldBe('entryValues().length', '7');
    shouldBeEqualToString('entryValues()[0]', '2012-11-20T06:00Z');
    shouldBeEqualToString('entryValues()[1]', '2012-11-20T12:00Z');
    shouldBeEqualToString('entryValues()[2]', '2012-11-20T18:00Z');
    shouldBeEqualToString('entryValues()[3]', '2012-11-21T00:00Z');
    shouldBeEqualToString('entryValues()[4]', '2012-11-21T06:00Z');
    shouldBeEqualToString('entryValues()[5]', '2012-11-21T12:00Z');
    shouldBeEqualToString('entryValues()[6]', '@openCalendarPicker');

    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
